+++ apps/api/apps/root/src/rest.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { PostModule } from './posts/post.module';
+
+@Module({
+  imports: [PostModule],
+})
+export class RestServiceModule {}