+++ packages/types/src/mongo/query.ts
@@ -1,0 +1,15 @@
+import { Callback, FilterQuery, ProjectionType, QueryOptions } from 'mongoose';
+
+export interface MongoQuery<T> {
+  filter?: FilterQuery<T>;
+  projection?: ProjectionType<T>;
+  options?: QueryOptions<T>;
+  callback?: Callback<T>;
+}
+
+export type Mquery<T> = [
+  filter?: FilterQuery<T>,
+  projection?: ProjectionType<T>,
+  options?: QueryOptions<T>,
+  callback?: Callback<T>,
+];