+++ packages/util/src/mongo/decorators/errorResponse.decorator.ts
@@ -1,0 +1,5 @@
+import { HttpStatus } from '@nestjs/common';
+
+export function ErrorResponse({ message }) {
+  return function (target: unknown, propertyKey: string, descriptor: PropertyDescriptor) {};
+}