+++ apps/web/CHANGELOG.md
@@ -3,9 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.0.1-next.1](https://personal-github.com/Tushar1998/ultimate-monorepo/compare/v0.0.1-next.0...v0.0.1-next.1) (2022-07-10)
+## [0.0.1-next.2](https://github.com/Tushar1998/ultimate-monorepo/compare/v0.0.1-next.1...v0.0.1-next.2) (2022-08-03)
 
-**Note:** Version bump only for package @ultimate/web
+**Note:** Version bump only for package @tushar1998/web
+
+
+
+
+
+## [0.0.1-next.1](https://personal-github.com/tushar1998/ultimate-monorepo/compare/v0.0.1-next.0...v0.0.1-next.1) (2022-07-10)
+
+**Note:** Version bump only for package @tushar1998/web
 
 
 
@@ -13,4 +21,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## 0.0.1-next.0 (2022-07-10)
 
-**Note:** Version bump only for package @ultimate/web
+**Note:** Version bump only for package @tushar1998/web